+++ src/gauge/shift_fields.jl
@@ -9,9 +9,15 @@ shift_fields(l::Gaussian, a::AbstractArray) = Gaussian(; θ = l.θ .+ a, l.γ)
 shift_fields(l::ReLU, a::AbstractArray) = ReLU(; θ = l.θ .+ a, l.γ)
 shift_fields(l::dReLU, a::AbstractArray) = dReLU(; θp = l.θp .+ a, θn = l.θn .+ a, l.γp, l.γn)
 shift_fields(l::pReLU, a::AbstractArray) = pReLU(; θ = l.θ .+ a, l.γ, l.Δ, l.η)
-shift_fields(l::xReLU, a::AbstractArray) = xReLU(; θ = l.θ .+ a, l.γ, l.Δ, l.ξ)
+function shift_fields(l::xReLU{N,A,FixGamma}, a::AbstractArray) where {N,A,FixGamma}
+    if FixGamma
+        return xReLU(; θ = l.θ .+ a, l.Δ, l.ξ)
+    else
+        return xReLU(; θ = l.θ .+ a, l.γ, l.Δ, l.ξ)
+    end
+end
 
-function shift_fields!(l::Union{Binary,Spin,Potts,PottsGumbel,Gaussian,ReLU,pReLU,xReLU,nsReLU}, a::AbstractArray)
+function shift_fields!(l::Union{Binary,Spin,Potts,PottsGumbel,Gaussian,ReLU,pReLU,xReLU}, a::AbstractArray)
     l.θ .+= a
     return l
 end