+++ src/standardized.jl
@@ -167,7 +167,7 @@ function ∂regularize_add_visible_offset!(∂::∂RBM, visible_regularization::
     ∂.w .-= (visible_regularization[1, ..] + visible_regularization[2, ..]) .* offset_h ./ scale_w
 end
 
-function ∂regularize_add_visible_offset!(∂::∂RBM, visible_regularization::AbstractArray, offset_h::AbstractArray, scale_w::AbstractArray, ::Union{Binary,Spin,Potts,PottsGumbel,Gaussian,ReLU,xReLU,pReLU,nsReLU})
+function ∂regularize_add_visible_offset!(∂::∂RBM, visible_regularization::AbstractArray, offset_h::AbstractArray, scale_w::AbstractArray, ::Union{Binary,Spin,Potts,PottsGumbel,Gaussian,ReLU,xReLU,pReLU})
     ∂.w .-= visible_regularization[1, ..] .* offset_h ./ scale_w
 end